Add doxygen-generated documentation support
Now Lib.mak handle a new $D directory, where generated documentation should
be built. A new "doc" phony rule is created to build all the documentation.
Makefiles can append targets to the $(doc) variable to be build when the
"doc" is used as a goal.

A very specific but reasonably useful pattern rule is added to build
doxygen based documentation. See the rule comment for more details.

The example sub-project "otherproj" now includes doxygen-generated
documentation as an example.