Add a function to include sub-directory's Build.mak
Including a Build.mak in a sub-directory can be very error prone, since the
$S variable should be set to the relative path from $T before including the
Build.mak and restored afterwards.

The function include_subdirs is added to simplify this task, freeing the
user from doing all the dirty work and avoiding bugs.

The "otherproj" pkg-config file generation is moved to a subdirectory to
test how the new function works with deeply nested Build.mak.