Support non-propagated target-specific variables
Now the compile and link functions support non-propagated target-specific
variables.  If a variable is named like the target with a suffix
".EXTRA_FLAGS", it value is added to the compile/link command.

A non-propagated linker can be used too, defining a variable with the name
of the target with the suffix ".LINKER".