Allow compiling from directories named "build"
If the absolute directory where the project has a component named "build"
(or whatever $(BUILD_DIR_NAME) is), the build fails because the whole
project directory is excluded from the obj directory tree replication.

This patch fixes this bug by adding the complete absolute path to the
project top-level directory ($T) to the excludes. An obsolete comment is
removed too.