Split configuration into project-wide and local
If Toplevel.mak and Makeit.mak should not be edited by the user, and
Config.mak is supposed to be a local configuration file, we are missing a
place to have project-wide configuration, like default compiler flags and
flavors. This patch make Config.mak a project-wide configuration file and
adds Config.local.mak as a local user-wide configuration, to tweak local
parameters as colors, or default compiler to use.